backend/pyproject.toml
[tool.poetry]
name = "star-repo-search-backend"
version = "0.1.0"
description = "Backend for GitHub starred repositories search"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 